core-processor/requirements.txt
librosa>=0.10
numpy
matplotlib
pygame-ce
soundfile
scipy
av
demucs

basic-pitch
